+++ Tests/AudioKitTests/Node Tests/Generator Tests/AudioPlayer2Tests.swift
@@ -3,14 +3,7 @@ import AVFoundation
 import CAudioKit
 import XCTest
 
-class AudioPlayer2Tests: XCTestCase {
-    // C4 - C5
-    let chromaticScale: [AUValue] = [261.63, 277.18, 293.66, 311.13, 329.63,
-                                     349.23, 369.99, 392, 415.3, 440,
-                                     466.16, 493.88] // , 523.25
-
-    private static var tempFiles = [URL]()
-
+class AudioPlayer2Tests: AudioFileTestCase {
     // for waiting in the background for realtime testing
     func wait(for interval: TimeInterval) {
         let delayExpectation = XCTestExpectation(description: "delayExpectation")
@@ -18,60 +11,6 @@ class AudioPlayer2Tests: XCTestCase {
             delayExpectation.fulfill()
         }
         wait(for: [delayExpectation], timeout: interval + 1)
-    }
-
-    func generateTestFile(named name: String = "_io_audiokit_AudioPlayer2Tests_temp",
-                          ofDuration duration: TimeInterval = 2,
-                          frequencies: [AUValue]? = nil) -> URL? {
-        let frequencies = frequencies ?? chromaticScale
-        guard frequencies.count > 0 else { return nil }
-
-        let pitchDuration = AUValue(duration) / AUValue(frequencies.count)
-
-        Log("duration", duration, "pitchDuration", pitchDuration)
-
-        let osc = Oscillator(waveform: Table(.square))
-        let engine = AudioEngine()
-        engine.output = osc
-
-        let url = FileManager.default.temporaryDirectory.appendingPathComponent("\(name)-\(AudioPlayer2Tests.tempFiles.count).aiff")
-        try? FileManager.default.removeItem(at: url)
-
-        var audioFormatSettings = Settings.audioFormat.settings
-        audioFormatSettings["AVLinearPCMIsNonInterleaved"] = false
-        guard let file = try? AVAudioFile(forWriting: url,
-                                          settings: audioFormatSettings) else {
-            return nil
-        }
-
-        var startTime: AUValue = 0
-        var notes = [AutomationEvent]()
-        for pitch in frequencies {
-            notes.append(AutomationEvent(targetValue: pitch, startTime: startTime, rampDuration: 0))
-            startTime += pitchDuration
-        }
-
-        let zero = [AutomationEvent(targetValue: 0,
-                                    startTime: 0,
-                                    rampDuration: 0)]
-        let fadeIn = [AutomationEvent(targetValue: 1,
-                                      startTime: 0,
-                                      rampDuration: pitchDuration)]
-        let fadeOut = [AutomationEvent(targetValue: 0,
-                                       startTime: AUValue(duration) - pitchDuration,
-                                       rampDuration: pitchDuration)]
-
-        Log(name, "duration", duration, "notes will play at", notes.map { $0.startTime })
-
-        try? engine.avEngine.render(to: file, duration: duration, prerender: {
-            osc.start()
-            osc.$amplitude.automate(events: zero + fadeIn + fadeOut)
-            osc.$frequency.automate(events: notes)
-        })
-        Log("rendered test file to \(url)")
-
-        AudioPlayer2Tests.tempFiles.append(url)
-        return url
     }
 
     func createPlayer(duration: TimeInterval,
@@ -84,19 +23,12 @@ class AudioPlayer2Tests: XCTestCase {
             return nil
         }
 
-        guard let player = AudioPlayer(url: url, buffered: buffered) else {
+        guard let player = AudioPlayer(url: url,
+                                       buffered: buffered) else {
             return nil
         }
         player.volume = 0.1
         return player
-    }
-
-    func cleanup() {
-        for url in AudioPlayer2Tests.tempFiles {
-            Log("removeItem", url)
-
-            try? FileManager.default.removeItem(at: url)
-        }
     }
 }
 
@@ -176,7 +108,6 @@ extension AudioPlayer2Tests {
     }
 
     func testAVDynamicConnection() {
-
         guard let url = generateTestFile(ofDuration: 2,
                                          frequencies: [220, 440]),
             let buffer = try? AVAudioPCMBuffer(url: url) else {
@@ -206,45 +137,45 @@ extension AudioPlayer2Tests {
         // Start here and test passes.
         try! engine.start()
 
-        //player.play()
-        //sleep(6)
+        // player.play()
+        // sleep(6)
     }
 
     func testPlayerConnectionWithMixer() {
-        return
-        let engine = AudioEngine()
-        let outputMixer = Mixer()
-        guard let player = createPlayer(duration: 1) else {
-            XCTFail("Failed to create AudioPlayer")
-            return
-        }
-        outputMixer.addInput(player)
-        engine.output = outputMixer
-        let audio = engine.startTest(totalDuration: 2.0)
-
-        player.play()
-
-        audio.append(engine.render(duration: 1.0))
-
-        guard let player2 = createPlayer(duration: 1) else {
-            XCTFail("Failed to create AudioPlayer")
-            return
-        }
-        let localMixer = Mixer()
-
-        localMixer.addInput(player2)
-        outputMixer.addInput(localMixer)
-
-        player2.play()
-        audio.append(engine.render(duration: 1.0))
+//        return
+//        let engine = AudioEngine()
+//        let outputMixer = Mixer()
+//        guard let player = createPlayer(duration: 1) else {
+//            XCTFail("Failed to create AudioPlayer")
+//            return
+//        }
+//        outputMixer.addInput(player)
+//        engine.output = outputMixer
+//        let audio = engine.startTest(totalDuration: 2.0)
+//
+//        player.play()
+//
+//        audio.append(engine.render(duration: 1.0))
+//
+//        guard let player2 = createPlayer(duration: 1) else {
+//            XCTFail("Failed to create AudioPlayer")
+//            return
+//        }
+//        let localMixer = Mixer()
+//
+//        localMixer.addInput(player2)
+//        outputMixer.addInput(localMixer)
+//
+//        player2.play()
+//        audio.append(engine.render(duration: 1.0))
 
 //        testMD5(audio)
         // audio.audition()
     }
 }
 
-extension AudioPlayer2Tests {
-    func testZZZRemoveTempFiles() {
-        cleanup()
-    }
-}
+// extension AudioPlayer2Tests {
+//    func testZZZRemoveTempFiles() {
+//        cleanup()
+//    }
+// }
